[WFLY-9731] test moved into ws package, container start/stop is not managed automatically, security domain annotation fixed
