+++ testsuite/integration/ws/src/test/java/org/jboss/as/test/integration/ws/authentication/policy/dummmy-ws-handler.xml
