+++ testsuite/integration/ws/src/test/java/org/jboss/as/test/integration/ws/authentication/policy/resources/META-INF/jboss-webservices.xml
