+++ testsuite/integration/ws/src/test/resources/standard-jaxws-client-config.xml
